+++ NEWS.md
@@ -1,6 +1,18 @@
 
 # vctrs (development version)
 
+* New performant `data_frame()` constructor for creating data frames in a way
+  that follows tidyverse semantics. Among other things, inputs are recycled
+  using tidyverse recycling rules, strings are never converted to factors,
+  list-columns are easier to create, and unnamed data frame input is
+  automatically spliced.
+
+* New `df_list()` for safely and consistently constructing the data structure
+  underlying a data frame, a named list of equal-length vectors. It is useful
+  in combination with `new_data_frame()` for creating user-friendly
+  constructors for data frame subclasses that use the tidyverse rules for
+  recycling and determining types.
+
 * Fixed performance issue with `vec_order()` on classed vectors which
   affected `dplyr::group_by()` (tidyverse/dplyr#5423).